requirements.txt
Flask==1.1.2
google-cloud-datastore==2.1.0
cirq==0.9.1
ply==3.11
